chatbot/requirements.txt
Django==5.2.5
djangorestframework==3.14.0
python-dotenv==1.0.0
qdrant-client==1.7.0
sentence-transformers==3.0.1
google-generativeai==0.3.0
langdetect==1.0.9
transformers==4.35.2
torch==2.9.0
PyMuPDF==1.23.8
python-docx==0.8.11
numpy==1.26.1
scikit-learn==1.3.2
requests==2.31.0
pymongo==4.6.1
dnspython==2.4.2
